backend/requirements_dev.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-extensions==3.0.3
django-filter==2.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.1.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
